backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
langchain==0.2.6
langchain-groq==0.1.6
langchain-community==0.2.6
langchain-huggingface==0.0.3
chromadb==0.5.3
pymupdf==1.24.5
python-dotenv==1.0.1
sentence-transformers==3.0.1
pydantic==2.7.4